src/main/java/com/example/kursach_proba/service/UserService.java
package com.example.kursach_proba.service;

import com.example.kursach_proba.persistence.entity.User;
import com.example.kursach_proba.persistence.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;


@Service
public class UserService {

    @Autowired
    UserRepository userRepository;

    public User save(User user) {
        return userRepository.save(user);
    }

    public User findByUsername(String username) {
        return userRepository.findByUsername(username);
    }
}
